refactor: moves DropdownFilterAjax to deprecated FE-2007
