+++ src/__deprecated__/components/dropdown-filter-ajax/dropdown-filter-ajax.stories.js
@@ -8,13 +8,13 @@ import {
 } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
 import { State, Store } from '@sambego/storybook-state';
-import { classicThemeSelector } from '../../../.storybook/theme-selectors';
-import OptionsHelper from '../../utils/helpers/options-helper';
+import { classicThemeSelector } from '../../../../.storybook/theme-selectors';
+import OptionsHelper from '../../../utils/helpers/options-helper';
 import notes from './documentation/notes.md';
 import Info from './documentation/Info';
 import DropdownFilterAjax from './dropdown-filter-ajax';
-import { enableMock } from '../../../demo/xhr-mock';
-import getDocGenInfo from '../../utils/helpers/docgen-info';
+import { enableMock } from '../../../../demo/xhr-mock';
+import getDocGenInfo from '../../../utils/helpers/docgen-info';
 
 DropdownFilterAjax.__docgenInfo = getDocGenInfo(
   require('./docgenInfo.json'),
