+++ src/components/inline-inputs/__definition__.js
@@ -1,7 +1,7 @@
 import InlineInputs from './';
 import TextboxDefinition from '../../__deprecated__/components/textbox/__definition__';
 import DecimalDefinition from '../../__deprecated__/components/decimal/__definition__';
-import DropdownFilterAjaxDefintion from './../dropdown-filter-ajax/__definition__';
+import DropdownFilterAjaxDefintion from '../../__deprecated__/components/dropdown-filter-ajax/__definition__';
 import Definition from '../../../demo/utils/definition';
 
 let definition = new Definition('inline-inputs', InlineInputs, {
